Refactored: HTTP to use Datadog::Contrib::Integration. (#556)
