+++ lib/ddtrace.rb
@@ -56,7 +56,7 @@ require 'ddtrace/contrib/faraday/patcher'
 require 'ddtrace/contrib/grape/patcher'
 require 'ddtrace/contrib/graphql/patcher'
 require 'ddtrace/contrib/grpc/patcher'
-require 'ddtrace/contrib/http/patcher'
+require 'ddtrace/contrib/http/integration'
 require 'ddtrace/contrib/integration'
 require 'ddtrace/contrib/mongodb/patcher'
 require 'ddtrace/contrib/mysql2/patcher'
